chore: move old query builder to deprecated folder
