+++ frontend/src/query/deprecated/Column/ColumnExpressionPicker.vue
@@ -118,7 +118,7 @@
 
 <script setup>
 import Code from '@/components/Controls/Code.vue'
-import ExpressionHelpDialog from '../ExpressionHelpDialog.vue'
+import ExpressionHelpDialog from '@/query/ExpressionHelpDialog.vue'
 import Tooltip from '@/components/Tooltip.vue'
 import { debounce } from 'frappe-ui'
 import ExpressionHelp from '@/components/ExpressionHelp.vue'
