+++ frontend/src/query/deprecated/Column/ColumnPicker.vue
@@ -24,8 +24,8 @@
 </template>
 
 <script setup>
-import SimpleColumnPicker from '@/query/Column/SimpleColumnPicker.vue'
-import ColumnExpressionPicker from '@/query/Column/ColumnExpressionPicker.vue'
+import SimpleColumnPicker from './SimpleColumnPicker.vue'
+import ColumnExpressionPicker from './ColumnExpressionPicker.vue'
 
 import { ref } from 'vue'
 defineEmits(['close'])
