+++ frontend/src/query/deprecated/Filter/CallExpression.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script setup>
-import ExpressionTerm from '@/query/Filter/ExpressionTerm.vue'
+import ExpressionTerm from './ExpressionTerm.vue'
 
 const props = defineProps({
 	expression: {
