+++ frontend/src/query/deprecated/Filter/FilterPanel.vue
@@ -36,8 +36,8 @@
 </template>
 
 <script setup>
-import LogicalExpression from '@/query/Filter/LogicalExpression.vue'
-import FilterPicker from '@/query/Filter/FilterPicker.vue'
+import LogicalExpression from './LogicalExpression.vue'
+import FilterPicker from './FilterPicker.vue'
 
 import { ref, inject, computed } from 'vue'
