+++ frontend/src/query/deprecated/Filter/LogicalExpression.vue
@@ -51,7 +51,7 @@
 
 <script setup>
 import { computed, nextTick, watch, ref, onMounted } from 'vue'
-import Expression from '@/query/Filter/Expression.vue'
+import Expression from './Expression.vue'
 
 defineEmits(['add-filter', 'edit-filter', 'remove-filter', 'toggle-operator'])
 const props = defineProps({
