+++ frontend/src/query/deprecated/Filter/SimpleFilterPicker.vue
